refactor: androidTest의 패키지 구조를 main 패키지 동일하게 적용
